added additional collision logic. added new models for map, including platform. These can now be created by making a config file that can be parsed with the XMLParser. core/assets/maps/map1.config is an example of such a file. any number of platforms and map objects can be added but only one map in each file.

+++ LibGDX/core/src/com/vormadal/turborocket/models/ammo/AmmoFactory.java
@@ -6,5 +6,5 @@ import com.vormadal.turborocket.WorldEntitiesController;
 
 public interface AmmoFactory  <T extends Ammo>{
 	T factory(Vector2 initialVel, Vector2 pos, Vector2 dir, WorldEntitiesController entitiesController);
-	int getAmmoCost();
+	float getAmmoCost();
 }
